+++ tests/deprecated_api/test_remove_1-6.py
@@ -21,7 +21,6 @@ from pytorch_lightning.plugins.training_type import DDPPlugin
 from pytorch_lightning.utilities.distributed import rank_zero_deprecation, rank_zero_warn
 from pytorch_lightning.utilities.model_helpers import is_overridden
 from pytorch_lightning.utilities.model_summary import ModelSummary
-from tests.deprecated_api import _soft_unimport_module
 from tests.helpers import BoringModel
 
 
@@ -116,10 +115,3 @@ def test_v1_6_0_deprecated_hpc_load(tmpdir):
     checkpoint_path = trainer.checkpoint_connector.get_max_ckpt_path_from_folder(str(tmpdir))
     with pytest.deprecated_call(match=r"`CheckpointConnector.hpc_load\(\)` was deprecated in v1.4"):
         trainer.checkpoint_connector.hpc_load(checkpoint_path)
-
-
-def test_v1_6_0_deprecated_device_dtype_mixin_import():
-
-    _soft_unimport_module("pytorch_lightning.utilities.device_dtype_mixin")
-    with pytest.deprecated_call(match="will be removed in v1.6"):
-        from pytorch_lightning.utilities.device_dtype_mixin import DeviceDtypeModuleMixin  # noqa: F401